Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
livekit-api==0.6.0
google-generativeai==0.8.3
faiss-cpu==1.8.0
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
redis==5.2.1
aiofiles==24.1.0
numpy==1.26.4
torch==2.1.0
transformers==4.36.0
